requirements.txt
attrs==21.4.0
Cerberus==1.3.4
click==8.0.4
coverage==6.2
#dataclasses==0.8
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
#psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.0.1
pytest-cov==3.0.0
six==1.16.0
SQLAlchemy==1.4.35
tomli==1.2.3
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
python-decouple==3.6
python-dotenv==0.20.0
alembic==1.8.1
